refactor: StrengthHeroInfoClient を StrengthHeroInfo に統合
- 両方がサーバーコンポーネントになったため統合
- StrengthHeroInfoClient.tsx と CSS を削除
- StrengthHeroInfo.tsx に表示ロジックを統合
- index.ts から不要な export を削除
