+++ src/features/strength/components/index.ts
@@ -1,5 +1,4 @@
 import StrengthHeroInfo from "./strength-hero-info/StrengthHeroInfo";
-import StrengthHeroInfoClient from "./strength-hero-info-client/StrengthHeroInfoClient";
 import StrengthHeroInfoSkeleton from "./strength-hero-info-skeleton/StrengthHeroInfoSkeleton";
 import StrengthLogInfo from "./strength-log-info/StrengthLogInfo";
 import StrengthLogInfoSkeleton from "./strength-log-info-skeleton/StrengthLogInfoSkeleton";
@@ -10,7 +9,6 @@ import StrengthTitleInfoSkeleton from "./strength-title-info-skeleton/StrengthTi
 
 export {
 	StrengthHeroInfo,
-	StrengthHeroInfoClient,
 	StrengthHeroInfoSkeleton,
 	StrengthLogInfo,
 	StrengthLogInfoSkeleton,
