+++ src/features/strength/components/strength-hero-info/StrengthHeroInfo.module.css
@@ -30,6 +30,199 @@
 	}
 }
 
+.strength-hero-info-content-head {
+	display: block grid;
+	grid-template-columns: auto auto;
+	place-content: center;
+	gap: 25px;
+	background-color: var(--color-primary-dark-gray);
+	border: 2px solid var(--color-primary-white);
+	border-radius: var(--border-radius-small);
+	padding-block: 15px;
+	padding-inline: 15px;
+
+	@media (width < 1280px) {
+		padding-block: 40px;
+		padding-inline: 10px;
+	}
+
+	@media (width < 768px) {
+		grid-template-columns: auto;
+	}
+}
+
+.strength-hero-box {
+	width: 100%;
+	display: block grid;
+	grid-template-rows: auto auto;
+	place-items: center;
+	gap: 15px;
+}
+
+.strength-hero-icon-box {
+	width: 110px;
+	display: block grid;
+	place-items: center;
+	filter: drop-shadow(0px 2px 1.5px var(--color-primary-black));
+}
+
+.strength-hero-icon-image {
+	width: 100%;
+	height: auto;
+	display: block flow;
+}
+
+.strength-hero-name-box {
+	background-color: var(--color-primary-brown-dark);
+	border: 1.5mm ridge var(--color-border-primary);
+	padding-inline: 1.5px;
+	padding-block: 1.5px;
+	filter: drop-shadow(0px 2px 1.5px var(--color-primary-black));
+}
+
+.strength-hero-name {
+	font-size: 0.75rem;
+	font-weight: var(--font-weight-normal);
+	line-height: var(--leading-tight);
+	text-shadow: 1.5px 1.5px 0px var(--color-primary-black);
+	border: 1px solid var(--color-primary-brown-light);
+	background-color: var(--color-bg-panel);
+	align-items: center;
+	gap: 5px;
+	padding-block: 2px 3px;
+	padding-inline: 7px;
+	display: block flex;
+}
+
+.strength-hero-name-icon {
+	width: 13px;
+	height: auto;
+}
+
+.strength-level-info {
+	width: 100%;
+	display: block grid;
+	place-items: center;
+}
+
+.strength-level-display-box {
+	display: block grid;
+	background-color: var(--color-primary-brown-dark);
+	border: 2.5mm ridge var(--color-border-primary);
+	padding-block: 2.5px;
+	padding-inline: 2.5px;
+	filter: drop-shadow(0px 2px 1.5px var(--color-primary-black));
+}
+
+.strength-level-display {
+	display: block grid;
+	grid-template-columns: repeat(2, auto);
+	gap: 5px;
+	font-weight: var(--font-weight-x-bold);
+	line-height: var(--leading-none);
+	color: var(--color-primary-white);
+	padding-block: 10px;
+	padding-inline: 10px;
+	border: 2px solid var(--color-primary-brown-light);
+	background-color: var(--color-bg-panel);
+	text-shadow: 2px 2px 0px var(--color-primary-black);
+}
+
+.strength-level-display-text {
+	display: block grid;
+	place-self: end;
+	font-size: 1.5rem;
+}
+
+.strength-level-display-value {
+	height: 32px;
+	height: 100%;
+	display: block grid;
+	place-items: end;
+	font-size: 2rem;
+}
+
+.strength-level-progress-info {
+	display: block grid;
+}
+
+.strength-level-progress-box {
+	display: block grid;
+	gap: 12px;
+}
+
+.strength-level-progress-content {
+	height: 24px;
+	display: block flex;
+	align-items: end;
+	justify-content: space-between;
+	color: var(--color-primary-white);
+	padding-inline-end: 5px;
+}
+
+.strength-level-progress-text {
+	font-size: 1rem;
+	font-weight: var(--font-weight-normal);
+	line-height: var(--leading-none);
+	text-shadow: 1px 1px 0px var(--color-primary-black);
+}
+
+.strength-level-progress-info-remaining-articles {
+	display: block flex;
+	align-items: end;
+	gap: 5px;
+}
+
+.strength-level-progress-info-remaining-articles-em {
+	font-size: 1.5rem;
+	font-weight: var(--font-weight-x-bold);
+	line-height: var(--leading-none);
+	font-style: normal;
+	text-shadow: 0px 0px 10px var(--color-primary-yellow);
+	color: var(--color-primary-yellow);
+}
+
+.strength-level-progress-info-remaining-articles-unit {
+	font-size: 0.875rem;
+	font-weight: var(--font-weight-normal);
+	line-height: var(--leading-none);
+	text-shadow: 1px 1px 0px var(--color-primary-black);
+}
+
+.strength-level-progress-gauge-container {
+	display: block flex;
+	align-items: center;
+	gap: 8px;
+}
+
+.strength-level-progress-exp-icon {
+	width: 35px;
+	height: auto;
+	filter: drop-shadow(0 0 1.5px var(--color-primary-white));
+	object-fit: contain;
+	object-position: center;
+}
+
+.strength-level-progress-gauge-box {
+	width: 100%;
+	height: 20px;
+	padding: 3px 3.5px;
+	background-color: var(--bg-progress-gauge);
+	border: 2px solid var(--color-primary-white);
+	border-radius: var(--border-rounded);
+	overflow: hidden;
+	filter: drop-shadow(0px 1.5px 1px var(--color-primary-black));
+}
+
+.strength-level-progress-gauge {
+	height: 100%;
+	transition: width 0.6s ease-in-out;
+	background-color: var(--color-primary-yellow);
+	border-radius: 999em 10em 10em 999em;
+	box-shadow: 0 0 5px 0px var(--color-primary-yellow);
+}
+
+/* エラー表示 */
 .error-text {
 	padding: 20px;
 	font-size: 1rem;
